SP5 Cookbook: refactor primitives
