+++ src/ScottPlot5/ScottPlot5 Cookbook/Cookbook.cs
@@ -7,7 +7,7 @@ namespace ScottPlotCookbook;
 /// These functions are used to locate components internally.
 /// Consumers of the cookbook will interact with the <see cref="Query"/> class to get this information.
 /// </summary>
-public static class Cookbook
+internal static class Cookbook
 {
     public static readonly string OutputFolder = Path.Combine(GetRepoFolder(), "dev/www/cookbook/5.0");
