+++ src/ScottPlot5/ScottPlot5 Cookbook/CookbookTests.cs
@@ -1,5 +1,4 @@
-﻿using FluentAssertions;
-using ScottPlotCookbook.Recipes;
+﻿using ScottPlotCookbook.Recipes;
 
 namespace ScottPlotCookbook;
