+++ src/ScottPlot5/ScottPlot5 Cookbook/QueryTests.cs
@@ -1,6 +1,4 @@
-﻿using FluentAssertions;
-
-namespace ScottPlotCookbook;
+﻿namespace ScottPlotCookbook;
 
 internal class QueryTests
 {
