+++ src/ScottPlot5/ScottPlot5 Cookbook/Usings.cs
@@ -1,3 +1,4 @@
 ﻿global using NUnit.Framework;
 global using ScottPlot;
 global using System.Text;
+global using FluentAssertions;
